feat: Add hello.js file and handler function
This commit adds a new file hello.js to the project, which contains a handler function that responds with a hello message and some event data. The function also parses the event data if it is a string. This change is necessary to handle incoming requests and provide a response.